% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_medias.R, R/tabela_totais_uf.R
\name{tabela_totais_uf}
\alias{tabela_totais_uf}
\title{Tabela de totais por UF e categoria para variável da PNS}
\usage{
tabela_totais_uf(variaveis = c())

tabela_totais_uf(variaveis = c())
}
\arguments{
\item{variaveis}{Vetor com nome(s) da(s) variável(is) de interesse.}
}
\value{
Um data frame em formato wide com totais e medidas de erro por UF e categoria.

Data frame com totais, erro padrão, intervalo de confiança e coeficiente de variação por categoria e UF.
}
\description{
Calcula os totais, coeficientes de variação e intervalos de confiança por UF e categoria da primeira variável informada,
com base na PNS 2019.
}
\examples{
tabela_totais_uf(variaveis = c("P040"))
}
